Fix Home drawText crash and Insights pane collapse bug
- _DomainWeekChart: label_w was a float (w*0.4) - drawText requires int
  sizes, so painting the Home wellbeing chart raised TypeError every
  frame
- AppShell: _apply_compact_shell_layout pushed setSizes on every call;
  with setChildrenCollapsible(True), the 0-width Insights entry
  permanently collapsed the pane - un-draggable afterwards. Sizes now
  apply only when panel visibility changes, and the splitter no longer
  collapses children

+++ litebrowser/ui/shell/pages.py
@@ -139,7 +139,7 @@ class _DomainWeekChart(QWidget):
         top = 8
         row_h = max(18, (h - 16) // max(1, len(self._rows)))
         max_count = max(c for _d, c, _s in self._rows) or 1
-        label_w = min(150, w * 0.4)
+        label_w = int(min(150, w * 0.4))
         bar_x = label_w + 10
         bar_w_max = max(30, w - bar_x - 46)
         for i, (domain, count, scary) in enumerate(self._rows):
